+++ module/durations.js
@@ -18,7 +18,7 @@ export async function checkCombatDurations(combat) {
     const canvasActors = canvas.tokens.objects.children.map(t => t.actor);
     const combatActors = combat.turns.map(t => t.actor);
 
-    [...canvasActors, ...combatActors].forEach(a => actorSet.add(a));
+    [...canvasActors, ...combatActors].filter(a => a).forEach(a => actorSet.add(a));
     const effects = [...actorSet].flatMap(actor => actor.effects.filter(e => e.isTemporary && e.duration?.type === "turns"));
     const expiredEffects = effects.filter(effect => {
         const dur = effect.duration;